Refactor the entity descriptions to pave way for removing entity_helpers.
