+++ custom_components/linksys_velop/entity_helpers.py
@@ -64,8 +64,6 @@ from .const import (
     CONF_COORDINATOR,
     DOMAIN,
     ENTITY_SLUG,
-    SIGNAL_UPDATE_CHECK_FOR_UPDATES_STATUS,
-    SIGNAL_UPDATE_SPEEDTEST_STATUS,
 )
 from .data_update_coordinator import LinksysVelopDataUpdateCoordinator
 from .logger import VelopLogger
@@ -332,103 +330,6 @@ class LinksysVelopNodeSensorPolled(LinksysVelopSensorPolled, LinksysVelopNodeEnt
 
         LinksysVelopSensorPolled.__init__(self, coordinator)
         LinksysVelopNodeEntity.__init__(self)
-
-
-# region #-- buttons --#
-@dataclasses.dataclass
-class OptionalLinksysVelopDescription:
-    """Represent the optional attributes of the button description."""
-
-    press_action_arguments: Optional[dict] = dict
-
-
-@dataclasses.dataclass
-class RequiredLinksysVelopDescription:
-    """Represent the required attributes of the button description."""
-
-    press_action: str
-
-
-@dataclasses.dataclass
-class LinksysVelopButtonDescription(
-    OptionalLinksysVelopDescription,
-    ButtonEntityDescription,
-    RequiredLinksysVelopDescription
-):
-    """Describes button entity"""
-
-
-BUTTONS: tuple[LinksysVelopButtonDescription, ...] = (
-    LinksysVelopButtonDescription(
-        key="",
-        name="Check for Updates",
-        press_action="async_check_for_updates",
-        press_action_arguments={
-            "signal": SIGNAL_UPDATE_CHECK_FOR_UPDATES_STATUS
-        }
-    ),
-    LinksysVelopButtonDescription(
-        key="",
-        name="Start Speedtest",
-        press_action="async_start_speedtest",
-        press_action_arguments={
-            "signal": SIGNAL_UPDATE_SPEEDTEST_STATUS
-        }
-    ),
-)
-# endregion
-
-
-# region #-- base entities --#
-class LinksysVelopMeshEntityByDescription(Entity):
-    """"""
-
-    def __init__(self, mesh: Mesh, config_entry: ConfigEntry) -> None:
-        """"""
-
-        self._config = config_entry
-        self._mesh: Mesh = mesh
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return the device information of the entity."""
-
-        # noinspection HttpUrlsUsage
-        ret = DeviceInfo(**{
-            "configuration_url": f"http://{self._mesh.connected_node}",
-            "identifiers": {(DOMAIN, self._config.entry_id)},
-            "manufacturer": PYVELOP_AUTHOR,
-            "model": f"{PYVELOP_NAME} ({PYVELOP_VERSION})",
-            "name": "Mesh",
-            "sw_version": "",
-        })
-        return ret
-
-
-class LinksysVelopNodeEntityByDescription(Entity):
-    """"""
-
-    def __init__(self, node: Node, config_entry: ConfigEntry) -> None:
-        """"""
-
-        self._config = config_entry
-        self._node: Node = node
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return the device information of the entity."""
-
-        ret = DeviceInfo(**{
-            "hw_version": self._node.hardware_version,
-            "identifiers": {(DOMAIN, self._node.serial)},
-            "model": self._node.model,
-            "name": self._node.name,
-            "manufacturer": self._node.manufacturer,
-            "sw_version": self._node.firmware.get("version", ""),
-        })
-
-        return ret
-# endregion
 
 
 def entity_setup(
